Add solution for SQL problem #176 'Second Highest Salary'
A new problem has been solved and added to the SQL category - the solution to finding the 'Second Highest Salary'. This solution utilizes a subquery, LIMIT and OFFSET clause, and distinct salaries to find the second highest salary. The README file has been appropriately updated to list this new problem.