Add solution for Leetcode problem #177: "Nth Highest Salary"
Included in the commit is the SQL code that uses LIMIT, OFFSET and ORDER BY clauses to solve Leetcode problem #177 "Nth Highest Salary". The solution retrieves the Nth highest salary from the Employee table and its details are explained in a new added README. This new problem and solution has also been added to the main README file.