Add solution for 'Sum of all subset XOR totals' problem
This commit includes the solution for the Leetcode problem "1863. Sum of all subset XOR totals" implemented in C++. The solution uses a recursive approach to compute the XOR sum of all possible subsets of the input array, with a time complexity of O(2^n) and space complexity of O(n). A link to the solution is added to the README.