Added solution for calculating score of a string
This commit introduces the functionality to calculate the score of a string. The solution first calculates the ASCII values of each character in the string and its immediately following character, then it finds the absolute difference between these two values and adds that difference to a cumulative score. This algorithm is implemented in PHP.