Add c++ solution for "Score of a string" problem
A new solution has been added to the repository for the "Score of a string" problem from LeetCode. The approach involves calculating the output by summing up the absolute difference between the ASCII values of consecutive characters. The detailed problem description, approach, complexity analysis, and solution are provided in the README.md file. The solution code is in Solution.cpp.