Add php in-place string reverse solution and documentation
Added a new PHP solution file and README.md documentation for the problem of reversing a string in-place. The solution uses a two-pointer approach to swap characters from the start and the end of the string and in turn, reverse the string. A respective entry in the main README.md was also created.