Remove algorithm and database problem descriptions
The commit removes a large amount of Markdown files that contain descriptions of various algorithm and database problems. This includes details on the problem itself and some example inputs and outputs. Additionally, the SQL code for some of the database problems were also removed. The selection of files seems to be randomly chosen and across different problem categories without a specific pattern.