Added solution for appending characters to string to make subsequence
Added a new solution to the leetcode problem where characters need to be appended to a string to make it a subsequence to another string. Implemented in PHP, the solution uses a two-pointer approach to traverse both strings efficiently and determine the number and position of characters to append. The README file was updated to reflect this added solution.